+++ R/ObjectiveFSelect.R
@@ -11,7 +11,7 @@
 #' @template param_measures
 #' @template param_store_models
 #' @template param_check_values
-#' @template param_store_resample_results
+#' @template param_store_benchmark_result
 #'
 #' @export
 ObjectiveFSelect = R6Class("ObjectiveFSelect",
@@ -36,13 +36,17 @@ ObjectiveFSelect = R6Class("ObjectiveFSelect",
     #' @field store_models (`logical(1)`).
     store_models = NULL,
 
-    #' @field store_resample_results (`logical(1)`).
-    store_resample_results = NULL,
+    #' @field store_benchmark_result (`logical(1)`).
+    store_benchmark_result = NULL,
+
+    #' @field archive ([ArchiveFSelect]).
+    archive = NULL,
 
     #' @description
     #' Creates a new instance of this [R6][R6::R6Class] class.
     initialize = function(task, learner, resampling, measures,
-      store_models = FALSE, check_values = TRUE, store_resample_results = TRUE) {
+      check_values = TRUE, store_benchmark_result = TRUE,
+      store_models = FALSE) {
 
       self$task = assert_task(as_task(task, clone = TRUE))
       self$learner = assert_learner(as_learner(learner, clone = TRUE),
@@ -51,8 +55,8 @@ ObjectiveFSelect = R6Class("ObjectiveFSelect",
         clone = TRUE))
       self$measures = assert_measures(as_measures(measures, clone = TRUE),
         task = self$task, learner = self$learner)
+      self$store_benchmark_result = assert_logical(store_benchmark_result)
       self$store_models = assert_logical(store_models)
-      self$store_resample_results = assert_logical(store_resample_results)
       if (!resampling$is_instantiated) {
         self$resampling$instantiate(self$task)
       }
@@ -85,9 +89,14 @@ ObjectiveFSelect = R6Class("ObjectiveFSelect",
       aggr = bmr$aggregate(self$measures)
       y = map_chr(self$measures, "id")
 
-      if (self$store_resample_results) {
-        rr = map(seq_len(bmr$n_resample_results), function(i) bmr$resample_result(i))
-        cbind(aggr[, y, with = FALSE], resample_result = rr)
+      if (self$store_benchmark_result) {
+        self$archive$benchmark_result =
+          if (is.null(self$archive$benchmark_result)) {
+            self$archive$benchmark_result = bmr
+          } else {
+            self$archive$benchmark_result$combine(bmr)
+          }
+        cbind(aggr[, y, with = FALSE], uhash = bmr$uhashes)
       } else {
         aggr[, y, with = FALSE]
       }