+++ tests/testthat/test_ObjectiveFSelect.R
@@ -15,9 +15,9 @@ test_that("ObjectiveFSelect", {
 
   z = obj$eval_many(xss)
   expect_data_table(z, nrows = 2, ncols = 2)
-  expect_equal(z$resample_result[[1]]$task$feature_names, c("x1", "x3", "x4"))
-  expect_equal(z$resample_result[[2]]$task$feature_names, c("x2", "x3", "x4"))
-  expect_null(z$resample_result[[1]]$learners[[1]]$model)
+  expect_equal(obj$archive$benchmark_result$resample_result(1)$task$feature_names, c("x1", "x3", "x4"))
+  expect_equal(obj$archive$benchmark_result$resample_result(2)$task$feature_names, c("x2", "x3", "x4"))
+  expect_null(obj$archive$benchmark_result$resample_result(1)$learners[[1]]$model)
 })
 
 test_that("ObjectiveFSelect works with multiple measures", {
@@ -52,5 +52,7 @@ test_that("ObjectiveFSelect works with store_models", {
     list("x1" = FALSE, "x2" = TRUE, "x3" = TRUE, "x4" = TRUE))
 
   z = obj$eval_many(xss)
-  expect_class(z$resample_result[[1]]$learners[[1]]$model, "rpart")
+  expect_class(
+    obj$archive$benchmark_result$resample_result(1)$learners[[1]]$model,
+    "rpart")
 })