+++ NEWS.md
@@ -1,3 +1,9 @@
+# mlr3fselect 0.2.1.9000
+
+* `Archive` is `ArchiveFSelect` now which stores the benchmark result in
+  `$benchmark_result`. This change removed the resample results from the archive
+   but they can be still accessed via the benchmark result.
+
 # mlr3fselect 0.2.1
 
 * Warning message if external package for feature selection is not installed.